services/dfs/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic-settings
sqlalchemy
psycopg2-binary
alembic
python-dotenv
